added graph navigator factories
